login page dan fitur-fiturnya sudah bisa digunakan, memperbaiki html register page
