+++ src/main/java/com/example/YuRun/Login/LoginController.java
@@ -1,52 +1,71 @@
 package com.example.YuRun.Login;
 
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.util.List;
+
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.stereotype.Controller;
+import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+
+import jakarta.servlet.http.HttpSession;
 
 @Controller
 @RequestMapping("/login")
-public class LoginController{
+public class LoginController {
+
     @Autowired
-    private LoginRepository repo;
+    private JdbcTemplate jdbcTemplate;
 
     @GetMapping
-    public String login(){
+    public String login() {
         return "/Login/index";
     }
 
-    /*
+    public List<LoginUser> findPengguna(String email) {
+        String sql = "SELECT * FROM Users WHERE email = ?";
+        return jdbcTemplate.query(sql, this::mapToPengguna, email);
+    }
+
+    private LoginUser mapToPengguna(ResultSet resultSet, int rowNum) throws SQLException {
+        return new LoginUser(
+                resultSet.getString("name"),
+                resultSet.getString("email"),
+                resultSet.getString("password"),
+                resultSet.getString("isadmin"),
+                resultSet.getBoolean("status")
+        );
+    }
+
     @PostMapping
     public String validation(
-        @RequestParam("name") String email,
-        @RequestParam("password") String password,
-        Model model, HttpSession session) {
+            @RequestParam("email") String email,
+            @RequestParam("password") String password,
+            Model model, HttpSession session) {
 
         List<LoginUser> pengguna = findPengguna(email);
 
         if (pengguna.isEmpty()) {
+            model.addAttribute("error", "Email or password is wrong");
             return "Login/index";
         } else {
             String realPassword = pengguna.get(0).getPassword();
-            if (realPassword.equals(password)) {    //Password benar
-                // Simpan nama pengguna dalam session
+            if (realPassword.equals(password)) { // Password benar
                 session.setAttribute("username", pengguna.get(0).getName());
-                // Jika yang login adalah admin
-                // Cek bit isAdmin dari database
-                // Asumsi ada method getIsAdmin() yang mengembalikan nilai bit (1 atau 0)
-                // int isAdmin = pengguna.get(0).getIsAdmin(); 
-
-                if (isAdmin == 1) {  // Jika 1 berarti admin
-                    return "redirect:/admin/LandingPage";
-                } else {  // Jika 0 berarti user biasa
-                    return "redirect:/user";
-                }   
-            } 
-            else {  //Password salah
-                return "LoginUser/index";
+                if (pengguna.get(0).getIsadmin().equals("1")) { // Jika admin
+                    return "redirect:/";
+                } else { // User biasa
+                    return "redirect:/user/home";
+                }
+            } else { // Password salah
+                model.addAttribute("error", "Email or password is wrong");
+                return "Login/index";
             }
         }
     }
-    */
 }
