--- src/main/java/com/example/YuRun/Login/LoginRepository.java
@@ -1,7 +1,0 @@
-package com.example.YuRun.Login;
-
-import java.util.List;
-
-public interface LoginRepository {
-    List<LoginUser> findPengguna(String email);
-}
