+++ src/main/java/com/example/YuRun/Register/RegisterController.java
@@ -1,0 +1,11 @@
+package com.example.YuRun.Register;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+
+public class RegisterController {
+    @GetMapping("/register")
+    public String register(){
+        return "/Register/index";
+    }
+}
