+++ src/main/resources/application.properties
@@ -2,8 +2,8 @@ spring.application.name=YuRun
 
 # DATABASE CONFIGURATION
 spring.datasource.url=jdbc:postgresql://localhost:5432/YURUN
-spring.datasource.username=postgres
-spring.datasource.password=FernandoJr1
+spring.datasource.username=rayhanadjisantoso
+spring.datasource.password=rayhan123
 spring.datasource.driver-class-name=org.postgresql.Driver
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQLDialect
 spring.jpa.hibernate.ddl-auto=update
