+++ src/main/resources/static/css/Login/styles.css
@@ -4,8 +4,8 @@ body {
     min-height: 100vh;
     min-width: 100vw;
     display: flex;
-    align-items: center; /* Center secara vertikal jika dibutuhkan */
-    justify-content: center; /* Center secara horizontal jika dibutuhkan */
+    align-items: center;
+    justify-content: center;
     grid-template-rows: auto auto auto;
     grid-template-columns: auto;
     grid-template-areas: 
@@ -31,8 +31,8 @@ body {
 .form {
     grid-area: form;
     display: flex;
-    align-items: center; /* Center secara vertikal jika dibutuhkan */
-    justify-content: center; /* Center secara horizontal jika dibutuhkan */
+    align-items: center;
+    justify-content: center;
     grid-template-rows: auto auto auto;
     grid-template-columns: auto;
     grid-template-areas: 
@@ -49,43 +49,48 @@ body {
     grid-area: password;
 }
 .label {
-  color: var(--primary-primary900, #8a2100);
-  text-align: left;
-  font-family: "Roboto-Regular", sans-serif;
-  font-size: 16px;
-  line-height: 100%;
-  letter-spacing: 0.01em;
-  font-weight: 400;
-  position: relative;
-  display: flex;
-  align-items: center;
-  justify-content: flex-start;
-}
-.input {
-  background: var(--secondary-secondary100, #e7ebef);
-  border-radius: 12px;
-  border-style: solid;
-  border-color: var(--secondary-secondary100, #e7ebef);
-  border-width: 1px;
-  flex-shrink: 0;
-  width: 400px;
-  height: 48px;
-  position: relative;
-}
-.button {
-    grid-area: button;
-    background: var(--primary-primary, #ff5722);
-    border-radius: 12px;
-    padding: 16px 0px 16px 0px;
-    align-self: stretch;
-    flex-shrink: 0;
-    color: var(--secondary-secondary100, #e7ebef);
-    text-align: center;
-    font-family: "Poppins-Regular", sans-serif;
-    font-size: 20px;
+    color: var(--primary-primary900, #8a2100);
+    text-align: left;
+    font-family: "Roboto-Regular", sans-serif;
+    font-size: 16px;
     line-height: 100%;
     letter-spacing: 0.01em;
     font-weight: 400;
+    position: relative;
+    display: flex;
+    align-items: center;
+    justify-content: flex-start;
+    margin-bottom: 10px; /* Tambahkan jarak antara label dan input */
+}
+
+.input input {
+    background: var(--secondary-secondary100, #e7ebef);
+    border-radius: 12px;
+    border: 1px solid var(--secondary-secondary100, #e7ebef);
+    width: 350px;
+    height: 50px;
+    padding: 0 12px;
+    box-sizing: border-box;
+    font-size: 16px;
+    color: #333;
+    outline: none;
+}
+  
+.input input:focus {
+    border-color: var(--secondary-secondary200, #c3c9cf);
+    background: var(--secondary-secondary50, #ffffff);
+}
+  
+.button button {
+    background: var(--primary-primary, #ff5722);
+    border: 1px solid var(--primary, #333);
+    color: var(--primary, #333);
+    border-radius: 12px;
+    width: 350px;
+    height: 50px;
+    font-size: 16px;
+    cursor: pointer;
+    text-align: center;
 }
 
 .signup {
