+++ src/main/resources/static/css/Register/styles.css
@@ -1,0 +1,126 @@
+.register,
+.register * {
+  box-sizing: border-box;
+}
+.register {
+  background: linear-gradient(to left, #cfcece, #cfcece);
+  height: 1024px;
+  position: relative;
+  overflow: hidden;
+}
+.already-have-an-account-sign-in {
+  text-align: center;
+  font-family: "-", sans-serif;
+  font-size: 18px;
+  line-height: 160%;
+  letter-spacing: 0.01em;
+  font-weight: 400;
+  position: absolute;
+  left: 582px;
+  top: 731px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+.already-have-an-account-sign-in-span {
+  color: var(--secondary-secondary800, #637d92);
+  font-family: "Roboto-Regular", sans-serif;
+  font-size: 18px;
+  line-height: 160%;
+  letter-spacing: 0.01em;
+  font-weight: 400;
+}
+.already-have-an-account-sign-in-span2 {
+  color: var(--primary-primary900, #8a2100);
+  font-family: "Roboto-Bold", sans-serif;
+  font-size: 18px;
+  line-height: 160%;
+  letter-spacing: 0.01em;
+  font-weight: 700;
+  text-decoration: underline;
+}
+.fieldset {
+  display: flex;
+  flex-direction: column;
+  gap: 24px;
+  align-items: flex-end;
+  justify-content: center;
+  width: 600px;
+  position: absolute;
+  left: 418px;
+  top: 265px;
+}
+.input {
+  display: flex;
+  flex-direction: column;
+  gap: 8px;
+  align-items: flex-start;
+  justify-content: flex-start;
+  align-self: stretch;
+  flex-shrink: 0;
+  position: relative;
+}
+.label {
+  color: var(--primary-primary900, #8a2100);
+  text-align: left;
+  font-family: "Poppins-Regular", sans-serif;
+  font-size: 16px;
+  line-height: 100%;
+  letter-spacing: 0.01em;
+  font-weight: 400;
+  position: relative;
+  display: flex;
+  align-items: center;
+  justify-content: flex-start;
+}
+.input2 {
+  background: var(--secondary-secondary100, #e7ebef);
+  border-radius: 12px;
+  border-style: solid;
+  border-color: var(--secondary-secondary100, #e7ebef);
+  border-width: 1px;
+  flex-shrink: 0;
+  width: 600px;
+  height: 48px;
+  position: relative;
+}
+.primary-button {
+  background: var(--primary-primary, #ff5722);
+  border-radius: 12px;
+  padding: 16px 0px 16px 0px;
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  justify-content: space-between;
+  align-self: stretch;
+  flex-shrink: 0;
+  position: relative;
+}
+.sign-in {
+  color: var(--secondary-secondary100, #e7ebef);
+  text-align: center;
+  font-family: "Poppins-Regular", sans-serif;
+  font-size: 20px;
+  line-height: 100%;
+  letter-spacing: 0.01em;
+  font-weight: 400;
+  position: relative;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+.welcome-to-yu-run {
+  color: var(--primary-primary900, #8a2100);
+  text-align: center;
+  font-family: "Poppins-SemiBold", sans-serif;
+  font-size: 32px;
+  line-height: 100%;
+  letter-spacing: 0.01em;
+  font-weight: 600;
+  position: absolute;
+  left: 418px;
+  top: 203.5px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
