+++ src/main/resources/templates/Register/index.html
@@ -1,0 +1,49 @@
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Document</title>
+</head>
+<body>
+    <div class="register"
+    style="
+        background: url(register.png) center;
+        background-size: cover;
+        background-repeat: no-repeat;
+    "
+    >
+    <div class="already-have-an-account-sign-in">
+        <span>
+        <span class="already-have-an-account-sign-in-span">
+            Already have an account?
+        </span>
+        <span class="already-have-an-account-sign-in-span2">Sign in</span>
+        </span>
+    </div>
+    <div class="fieldset">
+        <div class="input">
+        <div class="label">Full Name</div>
+        <div class="input2"></div>
+        </div>
+        <div class="input">
+        <div class="label">Email</div>
+        <div class="input2"></div>
+        </div>
+        <div class="input">
+        <div class="label">Password</div>
+        <div class="input2"></div>
+        </div>
+        <div class="input">
+        <div class="label">Confirm Password</div>
+        <div class="input2"></div>
+        </div>
+        <div class="primary-button">
+        <div class="sign-in">Sign up</div>
+        </div>
+    </div>
+    <div class="welcome-to-yu-run">Welcome to YuRun!</div>
+    </div>
+
+</body>
+</html>
